Move Evictor to its own module (#36)
* Move Evictor to its own module

* Adding clarifying information for the Evictor module

* Fix formatting of use statements

* Update mc-oblivious-ram/src/evictor/mod.rs


* Update mc-oblivious-ram/src/evictor/mod.rs


* Switch to newer module file format/naming convention
